Resolved issues with uniform handling now supporting setting uniform array from C-style array and context error generation when setting uniforms by name that is not found in program. Removed some type cast warnings. Resolved cgv namespace issues with planar_view_interactor. Added support to debug event dispatching to GLView and found and resolved bug in traverser with this. Finalized overlay such that event dispatching works for volume_rendering sample of examples plugin.

+++ cgv/media/color.h
@@ -853,7 +853,7 @@ std::string to_hex(const color<T, RGB, am>& c) {
 	res.resize(2 * static_cast<size_t>(color<T, RGB, am>::nr_components) + 2, '0');
 	std::string buf = "00";
 
-	for(size_t i = 0; i < static_cast<size_t>(color<T, RGB, am>::nr_components); ++i) {
+	for(unsigned i = 0; i < static_cast<size_t>(color<T, RGB, am>::nr_components); ++i) {
 		uint8_t v = 0;
 		convert_color_component(c[i], v);
 		buf = cgv::utils::to_hex(v);
